+++ VacationMachine.Repository/Dtos/EmployeeData.cs
@@ -1,0 +1,9 @@
+﻿namespace VacationMachine.Repository.Dtos
+{
+    public class EmployeeData
+    {
+        public int EmployeeId { get; set; }
+        public string Status { get; set; }
+        public int DaysOff { get; set; }
+    }
+}
