+++ VacationMachine.Services/Email/EmailSender.cs
@@ -1,4 +1,4 @@
-﻿namespace VacationMachine
+﻿namespace VacationMachine.Services.Email
 {
     public class EmailSender : IEmailSender
     {
