+++ VacationMachine.Services/Email/IEmailSender.cs
@@ -1,4 +1,4 @@
-﻿namespace VacationMachine
+﻿namespace VacationMachine.Services.Email
 {
     public interface IEmailSender
     {
