+++ VacationMachine.Services/IVacationService.cs
@@ -1,0 +1,9 @@
+﻿using VacationMachine.Services.Models;
+
+namespace VacationMachine.Services
+{
+    public interface IVacationService
+    {
+        Result RequestPaidDaysOff(int days, long employeeId);
+    }
+}
