+++ VacationMachine/VacationMachine.sln
@@ -1,9 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32126.315
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationMachine", "VacationMachine.csproj", "{6456D6A8-A681-414F-90B5-CA3025A079EF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VacationMachine", "VacationMachine.csproj", "{6456D6A8-A681-414F-90B5-CA3025A079EF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationMachine.Services", "..\VacationMachine.Services\VacationMachine.Services.csproj", "{2FCAB772-7E5C-4B51-8318-0DC1E7280D83}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationMachine.Repository", "..\VacationMachine.Repository\VacationMachine.Repository.csproj", "{5F5EC1FE-A731-457E-AF7B-CB345A2339F2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{6456D6A8-A681-414F-90B5-CA3025A079EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6456D6A8-A681-414F-90B5-CA3025A079EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6456D6A8-A681-414F-90B5-CA3025A079EF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2FCAB772-7E5C-4B51-8318-0DC1E7280D83}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2FCAB772-7E5C-4B51-8318-0DC1E7280D83}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2FCAB772-7E5C-4B51-8318-0DC1E7280D83}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2FCAB772-7E5C-4B51-8318-0DC1E7280D83}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5F5EC1FE-A731-457E-AF7B-CB345A2339F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5F5EC1FE-A731-457E-AF7B-CB345A2339F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5F5EC1FE-A731-457E-AF7B-CB345A2339F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5F5EC1FE-A731-457E-AF7B-CB345A2339F2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
